items: port Item_GetFrames to TRX
