+++ src/tr1/game/items.c
@@ -22,55 +22,3 @@ void Item_Control(void)
 
     Carrier_AnimateDrops();
 }
-
-int32_t Item_GetFrames(const ITEM *item, ANIM_FRAME *frames[], int32_t *rate)
-{
-    const ANIM *const anim = Item_GetAnim(item);
-    if (anim->frame_ptr == nullptr) {
-        frames[0] = nullptr;
-        return 0;
-    }
-
-    const int32_t cur_frame_num = item->frame_num - anim->frame_base;
-    const int32_t last_frame_num = anim->frame_end - anim->frame_base;
-    const int32_t key_frame_span = anim->interpolation;
-    const int32_t first_key_frame_num = cur_frame_num / key_frame_span;
-    const int32_t second_key_frame_num = first_key_frame_num + 1;
-
-    frames[0] = &anim->frame_ptr[first_key_frame_num];
-    frames[1] = &anim->frame_ptr[second_key_frame_num];
-
-    const int32_t key_frame_shift = cur_frame_num % key_frame_span;
-    const int32_t numerator = key_frame_shift;
-    int32_t denominator = key_frame_span;
-    if (numerator && second_key_frame_num > anim->frame_end) {
-        denominator = anim->frame_end + key_frame_span - second_key_frame_num;
-    }
-
-    // OG
-    if (g_Config.rendering.fps == 30) {
-        *rate = denominator;
-        return numerator;
-    }
-
-    // interpolated
-    if (item != g_LaraItem
-        && (!item->active || item->status != IS_ACTIVE
-            || !Object_Get(item->object_id)->enable_interpolation)) {
-        *rate = denominator;
-        return numerator;
-    }
-
-    const double clock_ratio = Interpolation_GetWorldRate() - 0.5;
-    const double final =
-        (key_frame_shift + clock_ratio) / (double)key_frame_span;
-    const double interp_frame_num =
-        (first_key_frame_num * key_frame_span) + (final * key_frame_span);
-    if (interp_frame_num >= last_frame_num) {
-        *rate = denominator;
-        return numerator;
-    }
-
-    *rate = 10;
-    return final * 10;
-}
